Move ToolingApiSpec back to cc integ tests
We should prefer to use ToolingApiSpecification, which is the standard test infrastructure for running tests using the tooling API.
We should not be maintaining two separate frameworks for this.
