+++ platforms/core-configuration/configuration-cache/src/integTest/groovy/org/gradle/internal/cc/impl/isolated/FetchCustomModelForTargetProject.java
@@ -16,7 +16,7 @@
 
 package org.gradle.internal.cc.impl.isolated;
 
-import org.gradle.integtests.tooling.fixture.SomeToolingModel;
+import org.gradle.internal.cc.impl.fixtures.SomeToolingModel;
 import org.gradle.tooling.BuildAction;
 import org.gradle.tooling.BuildController;
 import org.gradle.tooling.model.gradle.BasicGradleProject;
