+++ platforms/core-configuration/configuration-cache/src/integTest/groovy/org/gradle/internal/cc/impl/promo/ConfigurationCacheToolingApiPromoIntegrationTest.groovy
@@ -17,9 +17,9 @@
 package org.gradle.internal.cc.impl.promo
 
 import org.gradle.integtests.fixtures.executer.GradleExecuter
-import org.gradle.integtests.tooling.fixture.SomeToolingModel
-import org.gradle.integtests.tooling.fixture.ToolingApiBackedGradleExecuter
-import org.gradle.integtests.tooling.fixture.ToolingApiSpec
+import org.gradle.internal.cc.impl.fixtures.SomeToolingModel
+import org.gradle.internal.cc.impl.fixtures.ToolingApiBackedGradleExecuter
+import org.gradle.internal.cc.impl.fixtures.ToolingApiSpec
 import org.gradle.internal.cc.impl.AbstractConfigurationCacheIntegrationTest
 
 import static org.gradle.integtests.fixtures.logging.ConfigurationCacheOutputNormalizer.PROMO_PREFIX
